Remove handled/semantic TODOs from TopLevelLibraryInvalidTests
